refactor(daemon): extract SandboxLease from the K8s driver
Moves the three lease/serialization maps out of K8sDriver into a new
SandboxLease: the per-Sandbox `busy` refcount, the per-Sandbox `modeQueue`
transition chain, and the per-agent `suspending` gate. The mechanics that
only touch them move with them — retain/release, queueMode/applyMode,
resolveResumeImage and forgetSandbox — and the driver keeps its full public
surface as thin delegation.

suspendIfIdle stays split along the seam it actually has: the driver looks
up the launch and hands the lease a callback for the launch/session state,
so the lease still publishes its gate synchronously with the `busy` read
before the first await, and ensureSandbox still waits the gate out before
adopting and before reading launches.
